Changed the baseType in parseExcelDataType method to Numeric from Float because the Round method does not work in Real data type.

+++ backend/src/services/UtilityService.ts
@@ -283,7 +283,7 @@ export class UtilityService {
         } else if (normalizedType === 'BOOLEAN') {
             return { baseType: 'BOOLEAN' };
         } else if (normalizedType === 'NUMBER') {
-            return { baseType: 'FLOAT' };
+            return { baseType: 'NUMERIC' };
         } else if (normalizedType === 'DATE') {
             return { baseType: 'DATE' };
         }
